Register key recorder events by default
Previously, the key recorder would not register its recorded event if there were no observations tied to the event. This resulted in the unexpected behavior that, after registering an event and restarting the app, the recorder would not display its recorded event. This change updates the recorder so that it registers the event regardless of whether or not any observations have been created.