Fix a hidden retain cycle
`proxy.observeRegistrationState(updateVisualAppearance)` actually captures `self`, so we need to explicitly specify a weak capture.

`keyDownMonitor` needed to be reworked because, when `deinit` called `keyDownMonitor.stop()`, if `keyDownMonitor` had not been created yet, it would try to create it and cause an over-release error. This had never shown up as a problem before due to the retain cycle never allowing `deinit` to be called.